+++ ensnano-flatscene/src/data.rs
@@ -591,12 +591,6 @@ impl<R: DesignReader> Data<R> {
         ret
     }
 
-    pub fn save_isometry(&mut self) {
-        for h in self.helices.iter() {
-            self.design.set_isometry(h.flat_id, h.isometry);
-        }
-    }
-
     pub fn is_xover_end(&self, nucl: &FlatNucl) -> Option<bool> {
         self.design.is_xover_end(&nucl.to_real())
     }