+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Use cursor icons to indicate that actions are being performed in the 2D/3D views
 - Make it possible to tilt the 3D view with shift + RClick, and the 2D view with alt + Left/Right
 - Make it possible to apply a symetry transformation to the 2d view by pressing ctrl/⌘ + Left/Up/Right/Down
+- Introduce a new DNA parameter: inclination
 
 ## 0.4.0
 - Remaps mouse buttons in the 2D view.